pyproject.toml
[tool.poetry]
name = "vcr-langchain"
version = "0.0.3"
description = "Record and replay LLM interactions for langchain"
authors = ["Amos Jun-yeung Ng <me@amos.ng>"]
readme = "README.md"
packages = [{include = "vcr_langchain"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
vcrpy = "^4.2.1"
langchain = "^0.0.66"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.11.4"
flake8 = "^6.0.0"
google-search-results = "^2.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
